Reorganise TCK into separate traits and files
