--- tck/src/main/resources/tck.conf
@@ -1,1 +1,0 @@
-akka.http.server.preview.enable-http2 = on
